# openapi.yaml
openapi: 3.0.1
info:
  title: API Gestion Mairie
  description: 'API REST pour la gestion des réservations de biens municipaux, des
    utilisateurs, des groupes et des pièces justificatives. Système de coordination
    pour la mairie.'
  contact:
    name: Équipe Mairie
    email: contact@mairie.fr
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Serveur de développement local
security:
- bearer-jwt: []
...
paths:
...
  /biens:
    get:
      tags:
      - Biens
      summary: Liste tous les biens
      description: Récupère la liste paginée des biens visibles avec filtres optionnels
      operationId: getAllBiens
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          default: creerLe
      - name: sortDir
        in: query
        required: false
        schema:
          type: string
          default: desc
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: categoryId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Liste des biens récupérée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageBienDTO'
      security:
      - bearer-jwt: []
    post:
      tags:
      - Biens
      summary: Crée un nouveau bien
      description: Crée un bien municipal (ADMIN uniquement)
      operationId: createBien
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BienDTO'
        required: true
      responses:
        '201':
          description: Bien créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BienDTO'
        '403':
          description: Accès refusé
      security:
      - bearer-jwt: []
  /biens/{id}:
    get:
      tags:
      - Biens
      summary: Récupère un bien par ID
      description: Retourne les détails d'un bien spécifique
      operationId: getBienById
      parameters:
      - name: id
        in: path
        description: ID du bien
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Bien trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BienDTO'
        '404':
          description: Bien non trouvé
      security:
      - bearer-jwt: []
    put:
      tags:
      - Biens
      summary: Met à jour un bien
      description: Modifie les informations d'un bien existant (ADMIN uniquement)
      operationId: updateBien
      parameters:
      - name: id
        in: path
        description: ID du bien
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BienDTO'
        required: true
      responses:
        '200':
          description: Bien mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BienDTO'
        '403':
          description: Accès refusé
        '404':
          description: Bien non trouvé
      security:
      - bearer-jwt: []
    delete:
      tags:
      - Biens
      summary: Supprime un bien
      description: Supprime définitivement un bien (ADMIN uniquement)
      operationId: deleteBien
      parameters:
      - name: id
        in: path
        description: ID du bien
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Bien supprimé avec succès
        '403':
          description: Accès refusé
        '404':
          description: Bien non trouvé
      security:
      - bearer-jwt: []
  /biens/{id}/reservations:
    get:
      tags:
      - Biens
      summary: Liste les réservations d'un bien
      description: Récupère toutes les réservations associées à un bien
      operationId: getReservationsByBien
      parameters:
      - name: id
        in: path
        description: ID du bien
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Liste des réservations récupérée
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationDTO'
        '404':
          description: Bien non trouvé
      security:
      - bearer-jwt: []
...
  /api/upload/piece-justificative/{reservationId}:
    post:
      tags:
      - Upload
      summary: Upload une pièce justificative
      description: Téléverse un fichier justificatif pour une réservation
      operationId: uploadPieceJustificative
      parameters:
      - name: reservationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Fichier uploadé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PieceJustificative'
        '400':
          description: Fichier invalide ou réservation non trouvée
      security:
      - bearer-jwt: []
...
components:
  schemas:
...
    UserDTO:
      type: object
      properties:
        id_user:
          type: string
          format: uuid
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        telephone:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - USER
          - MODERATEUR
          - TUTORED
          - SECRETARY
        statut:
          type: string
          enum:
          - ACTIF
          - INACTIF
          - SUSPENDU
          - PENDING
          - ACTIVE
          - REJECTED
    BienDTO:
      type: object
      properties:
        id_bien:
          type: string
          format: uuid
        nom:
          type: string
        description:
          type: string
        estVisible:
          type: boolean
        id_cat_bien:
          type: string
          format: uuid
        creerLe:
          type: string
          format: date-time
...
    ReservationDTO:
      type: object
      properties:
        id_reservation:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserDTO'
        bien:
          $ref: '#/components/schemas/BienDTO'
        date_debut:
          type: string
          format: date-time
        date_fin:
          type: string
          format: date-time
        est_valide:
          type: boolean
        statut_caution:
          type: string
          enum:
          - NON_REQUISE
          - EN_ATTENTE
          - VERSEE
          - RESTITUEE
...
    PieceJustificative:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        filepath:
          type: string
        uploaded_at:
          type: string
          format: date-time
        sent_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
...
    PageBienDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BienDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: 'Authentification JWT. Format: Bearer {token}'
      scheme: bearer
      bearerFormat: JWT